% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tip.R
\name{get_candidates}
\alias{get_candidates}
\title{Compute the Set of Similar Subjects}
\usage{
get_candidates(.i, .similarity_matrix, .num_candidates)
}
\arguments{
\item{.i}{The subject index (i.e. row index in a matrix for vector-variate data).}

\item{.similarity_matrix}{The matrix of similarity values.}

\item{.num_candidates}{The number of similar subjects extracted.}
}
\description{
Find the <.num_candidates> subjects that are most similar to subject .i.
}
