% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tip.R
\name{log_likelihood_fn}
\alias{log_likelihood_fn}
\title{Likelihood Models}
\usage{
log_likelihood_fn(
  .cluster_vector,
  .i,
  .prior_estimates_for_likelihood,
  .likelihood_model
)
}
\arguments{
\item{.cluster_vector}{A vector of cluster assignments after the invitation step.}

\item{.i}{The subject index that the likelihood is computed for.}

\item{.prior_estimates_for_likelihood}{A list of hyperparameters that are computed using
the data and are used in the likelihood function.}

\item{.likelihood_model}{The name of the likelihood model being used. Example: "NIW".}
}
\description{
Likelihood Models
}
