% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{make_invertible}
\alias{make_invertible}
\title{Automatic matrix inversion}
\usage{
make_invertible(.matrix, .tolerance = 0.01)
}
\arguments{
\item{.matrix}{The matrix that may or may not be invertible.}

\item{.tolerance}{If necessary, a small value that is ITERATIVELY added to
each diagonal element of a matrix until the matrix is invertible.}
}
\description{
Automatic matrix inversion
}
