% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setClass_setMethod_bcm.R
\name{plot,bcm,missing-method}
\alias{plot,bcm,missing-method}
\alias{plot}
\title{A function to return plots from a Bayesian Clustering Model (bcm) object}
\usage{
\S4method{plot}{bcm,missing}(x, y, ...)
}
\arguments{
\item{x}{A Bayesian Clustering Model (bcm) object}

\item{y}{Not used.}

\item{...}{Not used.}
}
\description{
A function to return plots from a Bayesian Clustering Model (bcm) object
}
