% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fn.R
\name{plot.tip}
\alias{plot.tip}
\title{Bayesian Clustering Visualizations}
\usage{
\method{plot}{tip}(bcm, .subject_names = vector(), .node_size = 2, .add_node_labels = FALSE)
}
\arguments{
\item{A}{Bayesian Clustering Model (bcm) object}
}
\value{
A ggplot2 histogram of the posterior number of clusters and a plot of the posterior number of clusters versus the Gibbs sampling iteration (i.e. a trace plot).
}
\description{
A function that produces plots of the Bayesian clustering output.
}
