% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tip.R
\name{prob_tip_i}
\alias{prob_tip_i}
\title{Table Invitation Prior Probability}
\usage{
prob_tip_i(.i, .similarity_matrix, .current_assignments, .num_clusters)
}
\arguments{
\item{.i}{The subject index (i.e. row index in a matrix for vector-variate data).}

\item{.similarity_matrix}{The matrix of similarity values.}

\item{.current_assignments}{The posterior assignments after the invitation step.}

\item{.num_clusters}{The number of clusters after the invitation step.}
}
\description{
Compute the prior probability that a subject belongs to a cluster.
}
