% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{recode}
\alias{recode}
\title{Recode a vector of integer values to start at 1}
\usage{
recode(.posterior_assignments)
}
\description{
A function to "recode" a vector. For example, 2, 3, 5, 6, 10, 2, 2, 2, 5 needs
to be recoded to 1, 2, 3, 4, 5, 1, 1, 1, 3. This function is used to ensure that the posterior
cluster assignments start at 1 (otherwise an error occurs).
}
