% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tip.R
\name{get_cpt_neighbors}
\alias{get_cpt_neighbors}
\title{Estimate The Number of Similar Subjects}
\usage{
get_cpt_neighbors(.distance_matrix)
}
\arguments{
\item{.distance_matrix}{A symmetric n x n matrix of distances}
}
\value{
A vector of integers where the (i)th integer corresponds to the number of subjects (observations) that are similar to the (i)th subject (observation).
}
\description{
Estimate the number of similar subjects \
using univariate multiple change point detection (i.e. binary segmentation).
}
