% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fn.R
\name{ggnet2_network_plot}
\alias{ggnet2_network_plot}
\title{Visualize the posterior similarity matrix (i.e. posterior probability matrix)}
\usage{
ggnet2_network_plot(
  .matrix_graph,
  .subject_names = vector(),
  .subject_class_names = NA,
  .class_colors,
  .class_shapes,
  .random_seed = 7,
  .node_size = 6,
  .add_node_labels = TRUE
)
}
\arguments{
\item{.matrix_graph}{A matrix M where each element Mij corresponds to the posterior
probability that subjects i and j are in the same cluster.}

\item{.subject_names}{An optional vector of subject names that will appear in the graph plot.}

\item{.subject_class_names}{An optional vector of class names that will influence each vertex's color and shape.}

\item{.class_colors}{An optional named vector of colors. The vector names are required to be
the unique .subject_class_names whereas the vector values are required to be the colors.}

\item{.class_shapes}{An optional named vector of shapes. The vector names are required to be
the unique .subject_class_names whereas the vector values areq required to be integers
(i.e. pch values like 15, 16, 17, and so on).}

\item{.random_seed}{The plot uses the random layout, so set a seed for reproducibility.}

\item{.node_size}{The size of each node (i.e. vertex) in the graph plot.}

\item{.add_node_labels}{TRUE or FALSE. Should individual node labels be added to each node (i.e. vertex) in the graph plot?}
}
\description{
A function that produces a ggnet2 network plot to visualize the posterior similarity matrix (i.e. the matrix of posterior probabilities).
}
