% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fn.R
\name{ggplot_line_point}
\alias{ggplot_line_point}
\title{Plot connected points using ggplot2.}
\usage{
ggplot_line_point(.x, .y, .xlab = "", .ylab = "")
}
\arguments{
\item{.x}{The variable on the horizontal axis.}

\item{.y}{The variable on the vertical axis.}

\item{.xlab}{The label on the horizontal axis.}

\item{.ylab}{The label on the vertical axis.}
}
\value{
A ggplot2 plot of .y versus .x with a label .xlab on the horizontal axis and label .ylab on the vertical axis.
}
\description{
A function to that produces a ggplot2 plot of .y versus .x
where points are added via geom_point() and the points are connected via geom_line().
}
\examples{
ggplot_line_point(.x = rnorm(10), .y = rnorm(10), .xlab = "x", .ylab = "y")
}
