% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fn.R
\name{ggplot_number_of_clusters_hist}
\alias{ggplot_number_of_clusters_hist}
\title{Plot the posterior distribution of the number of clusters.}
\usage{
ggplot_number_of_clusters_hist(.posterior_number_of_clusters)
}
\arguments{
\item{.posterior_number_of_clusters}{A vector of the number of clusters after posterior sampling
for each iteration in the Gibbs sampler.}
}
\description{
A function that produces a ggplot bar chart (i.e. geom_bar) that corresponds
to the posterior number of clusters. The vertical axis is normalized so that it displays
the posterior probability.
}
\examples{
num_clusters <- c(1,2,2,2,2,3,3,1,2,3,3,3,1,3)
ggplot_number_of_clusters_hist(.posterior_number_of_clusters = num_clusters)
}
