% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fn.R
\name{ggplot_number_of_clusters_trace}
\alias{ggplot_number_of_clusters_trace}
\title{Plot the trace plot of the posterior number of clusters.}
\usage{
ggplot_number_of_clusters_trace(.posterior_number_of_clusters)
}
\arguments{
\item{.posterior_number_of_clusters}{A vector of the number of clusters after posterior sampling
for each iteration in the Gibbs sampler.}
}
\description{
A function that produces a ggplot2 trace plot (i.e. geom_line)
with respect to the posterior number of clusters.
}
\examples{
num_clusters <- c(1,2,2,2,2,3,3,1,2,3,3,3,1,3)
ggplot_number_of_clusters_trace(.posterior_number_of_clusters = num_clusters)
}
