% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{partition_undirected_graph}
\alias{partition_undirected_graph}
\title{Partition an Undirected Graph}
\usage{
partition_undirected_graph(.graph_matrix, .num_components, .step_size)
}
\arguments{
\item{.graph_matrix}{A symmetric matrix that the analyst wishes to decompose into <.num_components> components.}

\item{.num_components}{The number of components that the analyst wishes to decompose <.graph_matrix> into.}

\item{.step_size}{The size of the update for the cutoff in the transformation max(0, .graph_matrix - cutoff)
where cutoff = cutoff + .step_size.}
}
\value{
A list with three elements.
}
\description{
A function that iteratively applies the transformation max(0, .graph_matrix - cutoff) until
there are <.num_components> graph components where cutoff = cutoff + .step_size.
}
