% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tip.R
\name{tip}
\alias{tip}
\title{Bayesian Clustering with the Table Invitation Prior}
\usage{
tip(
  .data = list(),
  .burn = 1000,
  .samples = 1000,
  .similarity_matrix,
  .init_num_neighbors,
  .likelihood_model = "CONSTANT",
  .subject_names = vector(),
  .num_cores = 1,
  .step_size = 0.001
)
}
\arguments{
\item{.data}{Data frame (vectors comprise a row in a data frame; NIW only) or a list of matrices (MNIW only) that the analyst wishes to cluster. Note: if .likelihood_model = "CONSTANT", then the .data argument has no effect.}

\item{.burn}{Non-negative integer; the number of burn-in iterations in the Gibbs sampler.}

\item{.samples}{Positive integer; the number of sampling iterations in the Gibbs sampler.}

\item{.similarity_matrix}{Matrix; an n x n marix of simlarity values.}

\item{.init_num_neighbors}{vector of positive integers; each (i)th positive integer corresponds to the estimate of the number of subjects that are similar to the (i)th subject.}

\item{.likelihood_model}{Character; the name of the likelihood model used to compute the posterior probabilities.}

\item{.subject_names}{Vector of characters; an optional vector of names for the individual subjects. This is useful for the plotting function.}

\item{.num_cores}{Positive integer; the number of cores to use.}

\item{.step_size}{Positive numeric. A parameter used to ensure matrices are invertible. A small number is iteratively added to a matrix diagonal (if necessary) until the matrix is invertible.}
}
\description{
Bayesian clustering with the Table Invitation Prior (TIP) and optional likelihood functions.
}
