% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tip.R
\name{tip}
\alias{tip}
\title{Bayesian Clustering with the Table Invitation Prior}
\usage{
tip(
  .data,
  .burn,
  .samples,
  .similarity_matrix,
  .init_num_neighbors,
  .likelihood_model = "CONSTANT",
  .subject_names = vector(),
  .num_cores = 1,
  .tolerance = 0.001
)
}
\arguments{
\item{.data}{A list of vectors, matrices, or tensors that the analyst wishes to cluster.}

\item{.burn}{The number of burn-in iterations in the Gibbs sampler.}

\item{.samples}{The number of sampling iterations in the Gibbs sampler.}

\item{.similarity_matrix}{An n x x marix of simlarity values.}

\item{.init_num_neighbors}{A vector of integers corresponding to the estimate of the number of subjects that are similar to the (i)th subject.}

\item{.likelihood_model}{The name of the likelihood model used to compute the posterior probabilities.}

\item{.subject_names}{An optional vector of names for the individual subjects.}

\item{.num_cores}{The number of cores to use.}

\item{.tolerance}{A parameter used to ensure matrices are invertible. A small number is iteratively added to a matrix diagonal (if necessary) until the matrix is invertible.}
}
\description{
Bayesian clustering with the Table Invitation Prior (TIP) and optional likelihood functions.
}
